stringify profile id
there was an annoying error where passport 0 wasn't authenticating properly but everything else was. this was caused by authjs turning the 0 into a falsy value somewhere.

pushing with caution because when i tested locally it was always authenticating with passport 0 even when i completed auth for other passports. i decided that this is fake and i'm pushing anyway but if it happens in prod too then i'm gonna revert